Available functions

Base.read

Base.read

Base.read

LabConnections.BeagleBone.assert_pwm_write

LabConnections.BeagleBone.bbparse

LabConnections.BeagleBone.bbparse

LabConnections.BeagleBone.closedev

LabConnections.BeagleBone.export_gpio

LabConnections.BeagleBone.export_led

LabConnections.BeagleBone.export_pwm

LabConnections.BeagleBone.getdev

LabConnections.BeagleBone.initdev

LabConnections.BeagleBone.listdev

LabConnections.BeagleBone.printdev

LabConnections.BeagleBone.run_server

LabConnections.BeagleBone.teardown

LabConnections.BeagleBone.teardown

LabConnections.BeagleBone.teardown

LabConnections.BeagleBone.to_string

LabConnections.BeagleBone.to_string

LabConnections.BeagleBone.to_string

LabConnections.BeagleBone.write!

LabConnections.BeagleBone.write!

LabConnections.BeagleBone.write!

Available functions

LabConnections.BeagleBone.run_server — Function.

run_server(port=2001; debug=false)

Run a server on port that listens for commands from computer Optional debug keyword disables blinking system leds.

Base.read — Function.

l = read(led::SysLED, debug::Bool=false)

Reads the current brightness value from the LED 'SysLED'.

Base.read — Function.

l = read(pwm::PWM, operation::Int32, debug::Bool=false)

Reads the current value from an operation on a GPIO.

Base.read — Function.

l = read(gpio::GPIO, operation::Int32, debug::Bool=false)

Reads the current value from an operation on a GPIO.

LabConnections.BeagleBone.assert_pwm_write — Method.

assert_pwm_write(operation::Int32, entry::String)

Assertsion for the PWM input data.

LabConnections.BeagleBone.bbparse — Method.

bbparse(cmd)

Parse and execute the command cmd.

LabConnections.BeagleBone.bbparse — Method.

bbparse(l::Tuple, sock)

Parse input on the form l=(iswrite, ndev, cmd1, cmd2, ..., cmdn) where if iswrite cmdi = (devname, id, val) and if not iswrite cmdi = (devname, id) and send back on socket (vals, timestamps).

LabConnections.BeagleBone.closedev — Method.

closedev(dev_name::String, i::Int32)

Closes down a currently active device of type 'dev_name' at index 'i' on the BeagleBone, and removes it from the dict of currently active devices.

LabConnections.BeagleBone.export_gpio — Method.

export_gpio(i::Int32, debug::Bool=false)

Export the GPIO file system, either for real-time or testing usecases.

LabConnections.BeagleBone.export_led — Function.

export_led(i::Int32, debug::Bool=false)

Exports a dummy filesystem for testing the LED implementation

LabConnections.BeagleBone.export_pwm — Method.

export_gpio(i::Int32, debug::Bool=false)

Export the GPIO file system, either for real-time or testing usecases.

LabConnections.BeagleBone.getdev — Method.

dev = getdev(dev_name::String, i::Int32)

Retrieves the active device of type dev_name at index 'i'.

LabConnections.BeagleBone.initdev — Method.

active_device = initdev(dev_name::String, i:Int32)

Initializes a new device of type 'devname' at index 'i' on the BeagleBone, and adds it to the dict of currently active devices. Returns the initialized device 'activedevice'.

LabConnections.BeagleBone.listdev — Method.

message = listdev()

Lists all the active devices as an insidence array for testing.

LabConnections.BeagleBone.printdev — Method.

message = printdev()

Prints all the active devices and writes out specifics of a single devices.

LabConnections.BeagleBone.teardown — Function.

teardown(led::SysLED, debug::Bool=false)

Closes all open filestreams for the SysLED 'led'.

LabConnections.BeagleBone.teardown — Function.

teardown(gpio::GPIO, debug::Bool=false)

Closes all open streams on the GPIO, and unexports it from the file system.

LabConnections.BeagleBone.teardown — Function.

teardown!(pwd::PWM)

Closes all open streams on the PWM, and unexports it from the file system

LabConnections.BeagleBone.to_string — Function.

to_string(led::SysLED, debug::Bool=false)

Generates a string representation of the GPIO device.

LabConnections.BeagleBone.to_string — Function.

to_string(gpio::GPIO, debug::Bool=false)

Generates a string representation of the GPIO device.

LabConnections.BeagleBone.to_string — Function.

to_string(pwm::PWM,, debug::Bool=false)

Generates a string representation of the GPIO device.

LabConnections.BeagleBone.write! — Function.

write!(gpio::GPIO, args::Tuple{Int32,String}, debug::Bool=false)

Writes an entry to an operation on a GPIO, of the form args = (operation, entry).

LabConnections.BeagleBone.write! — Function.

write!(led::SysLED, val::Bool, debug::Bool=false)

Turns the LED 'SysLed' on/off for val = true/false respectively.

LabConnections.BeagleBone.write! — Function.

write!(pwm::PWM, args::Tuple{Int32,String}, debug::Bool=false)

Writes an entry to an operation on the PWM, of the form args = (operation, entry).